requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
telethon==1.38.1
python-dotenv==1.0.1
werkzeug==3.0.3
psutil==5.9.8
requests==2.31.0
aiofiles==24.1.0
colorama==0.4.6
gunicorn
gevent
